Refactorings based on master merge
