+++ core/cas-server-core-services-registry/src/main/java/org/apereo/cas/services/ChainingServicesManager.java
@@ -166,6 +166,11 @@ public class ChainingServicesManager implements ServicesManager, DomainAwareServ
     }
 
     @Override
+    public int count() {
+        return this.getAllServices().size();
+    }
+
+    @Override
     public boolean supports(final Service service) {
         return findServicesManager(service).isPresent();
     }
